回退 'Pull Request !217 : smartperf-host告警清理'

+++ ide/src/trace/database/data-trafic/CpuDataReceiver.ts
@@ -123,7 +123,7 @@ function arrayBufferHandler(data: unknown, res: unknown[], transfer: boolean): v
   // @ts-ignore
   let cpu = new Uint8Array(transfer ? res.length : data.params.sharedArrayBuffers.cpu);
   // @ts-ignore
-  let argSetId = new Int32Array(transfer ? res.length : data.params.sharedArrayBuffers.argSetId);
+  let argSetId = new Int8Array(transfer ? res.length : data.params.sharedArrayBuffers.argSetId);
   // @ts-ignore
   let nofinish = new Uint8Array(transfer ? res.length : data.params.sharedArrayBuffers.nofinish);
   res.forEach((it, i) => {
